requirements.txt
ansible
docker
molecule
molecule-plugins[docker]
pytest-testinfra
gunicorn
